Rename module "resources" to "references" to avoid conflicts
Eventually, resources will be split into resource specifications and exact resource, and enriched with namespaces patterns and names.

For now, the goal is also to remove the conflicts with variable names and function arguments, where we had to use "resources_" aliasing.
