+++ kopf/reactor/registries.py
@@ -20,7 +20,7 @@ from typing import Any, Callable, Collection, Container, FrozenSet, Generic, Ite
                    List, Mapping, MutableMapping, Optional, Sequence, Set, TypeVar, Union, cast
 
 from kopf.reactor import causation, invocation
-from kopf.structs import callbacks, dicts, filters, handlers, resources as resources_
+from kopf.structs import callbacks, dicts, filters, handlers, references
 from kopf.utilities import piggybacking
 
 # We only type-check for known classes of handlers/callbacks, and ignore any custom subclasses.
@@ -116,7 +116,7 @@ class ResourceRegistry(
 
     def has_handlers(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
     ) -> bool:
         for handler in self._handlers:
             if _matches_resource(handler, resource):
@@ -140,13 +140,13 @@ class ResourceRegistry(
 
     def get_extra_fields(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
     ) -> Set[dicts.FieldPath]:
         return set(self.iter_extra_fields(resource=resource))
 
     def iter_extra_fields(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
     ) -> Iterator[dicts.FieldPath]:
         for handler in self._handlers:
             if _matches_resource(handler, resource):
@@ -174,13 +174,13 @@ class ResourceRegistry(
         warnings.warn("Direct usage of registries is deprecated.", DeprecationWarning)
         return len({h.resource for h in self._handlers if h.resource})
 
-    def __iter__(self) -> Iterator[resources_.Resource]:
+    def __iter__(self) -> Iterator[references.Resource]:
         warnings.warn("Direct usage of registries is deprecated.", DeprecationWarning)
         return iter({h.resource for h in self._handlers if h.resource})
 
     def __getitem__(
             self: "ResourceRegistry[CauseT, HandlerFnT, ResourceHandlerT]",
-            _: resources_.Resource,
+            _: references.Resource,
     ) -> "ResourceRegistry[CauseT, HandlerFnT, ResourceHandlerT]":
         warnings.warn("Direct usage of registries is deprecated.", DeprecationWarning)
         return self
@@ -195,7 +195,7 @@ class ResourceWatchingRegistry(ResourceRegistry[
             self,
             fn: callbacks.ResourceWatchingFn,
             *,
-            resource: resources_.Resource,
+            resource: references.Resource,
             id: Optional[str] = None,
             errors: Optional[handlers.ErrorsMode] = None,
             timeout: Optional[float] = None,
@@ -256,7 +256,7 @@ class ResourceChangingRegistry(ResourceRegistry[
             self,
             fn: callbacks.ResourceChangingFn,
             *,
-            resource: resources_.Resource,
+            resource: references.Resource,
             id: Optional[str] = None,
             reason: Optional[handlers.Reason] = None,
             event: Optional[str] = None,  # deprecated, use `reason`
@@ -319,7 +319,7 @@ class ResourceChangingRegistry(ResourceRegistry[
 
     def get_resource_handlers(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
     ) -> Sequence[handlers.ResourceChangingHandler]:
         found_handlers: List[handlers.ResourceChangingHandler] = []
         for handler in self._handlers:
@@ -348,7 +348,7 @@ class OperatorRegistry:
         self.resource_changing_handlers = ResourceChangingRegistry()
 
     @property
-    def resources(self) -> FrozenSet[resources_.Resource]:
+    def resources(self) -> FrozenSet[references.Resource]:
         """ All known resources in the registry. """
         return frozenset(
             {h.resource for h in self.resource_watching_handlers.get_all_handlers() if h.resource} |
@@ -399,7 +399,7 @@ class OperatorRegistry:
         warnings.warn("registry.register_resource_watching_handler() is deprecated; "
                       "use @kopf.on... decorators with registry= kwarg.",
                       DeprecationWarning)
-        resource = resources_.Resource(group, version, plural)
+        resource = references.Resource(group, version, plural)
         return self.resource_watching_handlers.register(
             fn=fn, id=id,
             resource=resource, labels=labels, annotations=annotations, when=when,
@@ -433,7 +433,7 @@ class OperatorRegistry:
         warnings.warn("registry.register_resource_changing_handler() is deprecated; "
                       "use @kopf.on... decorators with registry= kwarg.",
                       DeprecationWarning)
-        resource = resources_.Resource(group, version, plural)
+        resource = references.Resource(group, version, plural)
         return self.resource_changing_handlers.register(
             reason=reason, event=event, field=field, fn=fn, id=id,
             errors=errors, timeout=timeout, retries=retries, backoff=backoff, cooldown=cooldown,
@@ -451,7 +451,7 @@ class OperatorRegistry:
 
     def has_resource_watching_handlers(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
     ) -> bool:
         warnings.warn("registry.has_resource_watching_handlers() is deprecated; "
                       "please cease using the internal registries directly.",
@@ -460,7 +460,7 @@ class OperatorRegistry:
 
     def has_resource_changing_handlers(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
     ) -> bool:
         warnings.warn("registry.has_resource_changing_handlers() is deprecated; "
                       "please cease using the internal registries directly.",
@@ -531,7 +531,7 @@ class OperatorRegistry:
 
     def get_extra_fields(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
     ) -> Set[dicts.FieldPath]:
         warnings.warn("registry.get_extra_fields() is deprecated; "
                       "please cease using the internal registries directly.",
@@ -543,7 +543,7 @@ class OperatorRegistry:
 
     def iter_extra_fields(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
     ) -> Iterator[dicts.FieldPath]:
         warnings.warn("registry.iter_extra_fields() is deprecated; "
                       "please cease using the internal registries directly.",
@@ -554,7 +554,7 @@ class OperatorRegistry:
 
     def requires_finalizer(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
             cause: causation.ResourceCause,
     ) -> bool:
         """
@@ -697,7 +697,7 @@ def match(
 
 def _matches_resource(
         handler: handlers.ResourceHandler,
-        resource: resources_.Resource,
+        resource: references.Resource,
 ) -> bool:
     return (handler.resource is None or
             handler.resource == resource)
