+++ kopf/toolkits/legacy_registries.py
@@ -10,8 +10,7 @@ import warnings
 from typing import Any, Container, Iterator, Optional, Sequence, Union
 
 from kopf.reactor import causation, registries
-from kopf.structs import bodies, callbacks, dicts, filters, \
-                         handlers, patches, resources as resources_
+from kopf.structs import bodies, callbacks, dicts, filters, handlers, patches, references
 
 AnyCause = Union[causation.ResourceWatchingCause, causation.ResourceChangingCause]
 AnyHandler = Union[handlers.ResourceWatchingHandler, handlers.ResourceChangingHandler]
@@ -37,7 +36,7 @@ class BaseRegistry(metaclass=abc.ABCMeta):
     @abc.abstractmethod
     def get_event_handlers(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
             event: bodies.RawEvent,
     ) -> Sequence[handlers.ResourceWatchingHandler]:
         raise NotImplementedError
@@ -52,7 +51,7 @@ class BaseRegistry(metaclass=abc.ABCMeta):
     @abc.abstractmethod
     def iter_event_handlers(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
             event: bodies.RawEvent,
     ) -> Iterator[handlers.ResourceWatchingHandler]:
         raise NotImplementedError
@@ -126,7 +125,7 @@ class SimpleRegistry(BaseRegistry, registries.ResourceRegistry[
 
     def get_event_handlers(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
             event: bodies.RawEvent,
     ) -> Sequence[handlers.ResourceWatchingHandler]:
         warnings.warn("SimpleRegistry.get_event_handlers() is deprecated; "
@@ -146,7 +145,7 @@ class SimpleRegistry(BaseRegistry, registries.ResourceRegistry[
 
     def iter_event_handlers(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
             event: bodies.RawEvent,
     ) -> Iterator[handlers.ResourceWatchingHandler]:
         warnings.warn("SimpleRegistry.iter_event_handlers() is deprecated; "
@@ -211,7 +210,7 @@ class GlobalRegistry(BaseRegistry, registries.OperatorRegistry):
 
     def get_event_handlers(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
             event: bodies.RawEvent,
     ) -> Sequence[handlers.ResourceWatchingHandler]:
         warnings.warn("GlobalRegistry.get_event_handlers() is deprecated; "
@@ -231,7 +230,7 @@ class GlobalRegistry(BaseRegistry, registries.OperatorRegistry):
 
     def iter_event_handlers(
             self,
-            resource: resources_.Resource,
+            resource: references.Resource,
             event: bodies.RawEvent,
     ) -> Iterator[handlers.ResourceWatchingHandler]:
         warnings.warn("GlobalRegistry.iter_event_handlers() is deprecated; "
@@ -255,7 +254,7 @@ class SmartGlobalRegistry(registries.SmartOperatorRegistry, GlobalRegistry):
 
 
 def _create_watching_cause(
-        resource: resources_.Resource,
+        resource: references.Resource,
         event: bodies.RawEvent,
 ) -> causation.ResourceWatchingCause:
     return causation.detect_resource_watching_cause(
