+++ tests/registries/legacy-1/test_legacy1_requires_finalizer.py
@@ -3,7 +3,7 @@ import pytest
 import kopf
 from kopf import GlobalRegistry
 from kopf.structs.filters import MetaFilterToken
-from kopf.structs.resources import Resource
+from kopf.structs.references import Resource
 
 OBJECT_BODY = {
     'apiVersion': 'group/version',
