+++ tests/registries/legacy-2/test_legacy2_decorators.py
@@ -5,7 +5,7 @@ from kopf.reactor.handling import handler_var, subregistry_var
 from kopf.reactor.invocation import context
 from kopf.reactor.registries import OperatorRegistry, ResourceChangingRegistry
 from kopf.structs.handlers import HANDLER_REASONS, Activity, ErrorsMode, Reason
-from kopf.structs.resources import Resource
+from kopf.structs.references import Resource
 
 
 def test_on_startup_minimal():
