+++ tests/registries/legacy-2/test_legacy2_requires_finalizer.py
@@ -3,7 +3,7 @@ import pytest
 import kopf
 from kopf.reactor.registries import OperatorRegistry
 from kopf.structs.filters import MetaFilterToken
-from kopf.structs.resources import Resource
+from kopf.structs.references import Resource
 
 OBJECT_BODY = {
     'apiVersion': 'group/version',
